refactor(NcAppNavigationSpacer): migrate to Typescript
