+++ src/components/NcAppNavigationSpacer/NcAppNavigationSpacer.vue
@@ -3,15 +3,13 @@
   - SPDX-License-Identifier: AGPL-3.0-or-later
 -->
 
+<script setup lang="ts">
+// nothing
+</script>
+
 <template>
 	<li class="app-navigation-spacer" />
 </template>
-
-<script>
-export default {
-	name: 'NcAppNavigationSpacer',
-}
-</script>
 
 <style scoped>
 	.app-navigation-spacer {
