+++ src/components/index.ts
@@ -27,7 +27,7 @@ export { default as NcAppNavigationNew } from './NcAppNavigationNew/index.js'
 export { default as NcAppNavigationNewItem } from './NcAppNavigationNewItem/index.js'
 export { default as NcAppNavigationSearch } from './NcAppNavigationSearch/index.js'
 export { default as NcAppNavigationSettings } from './NcAppNavigationSettings/index.js'
-export { default as NcAppNavigationSpacer } from './NcAppNavigationSpacer/index.js'
+export { default as NcAppNavigationSpacer } from './NcAppNavigationSpacer/index.ts'
 export { default as NcAppSettingsDialog } from './NcAppSettingsDialog/index.js'
 export { default as NcAppSettingsSection } from './NcAppSettingsSection/index.js'
 export { default as NcAppSidebar } from './NcAppSidebar/index.js'
